Check delete Job and Company, add Company Profile

+++ app/Http/Controllers/Admin/AdminCompanyIndustryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\CompanyIndustry;
+use App\Models\Company;
 
 class AdminCompanyIndustryController extends Controller
 {
@@ -54,6 +55,10 @@ class AdminCompanyIndustryController extends Controller
     
     public function delete($id)
     {
+        $check = Company::where('company_industry_id', $id)->count();
+        if($check > 0) {
+            return redirect()->back()->with('error', 'You can not delete this item, because this is used in another place.');
+        }
         CompanyIndustry::where('id', $id)->delete();
         return redirect()->route('admin_company_industry')->with('success', 'Data is deleted successfully.');
     }
