+++ app/Http/Controllers/Admin/AdminCompanySizeController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\CompanySize;
+use App\Models\Company;
 
 class AdminCompanySizeController extends Controller
 {
@@ -54,6 +55,10 @@ class AdminCompanySizeController extends Controller
     
     public function delete($id)
     {
+        $check = Company::where('company_size_id', $id)->count();
+        if($check > 0) {
+            return redirect()->back()->with('error', 'You can not delete this item, because this is used in another place.');
+        }
         CompanySize::where('id', $id)->delete();
         return redirect()->route('admin_company_size')->with('success', 'Data is deleted successfully.');
     }
