+++ app/Http/Controllers/Admin/AdminJobCategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\JobCategory;
+use App\Models\Job;
 
 class AdminJobCategoryController extends Controller
 {
@@ -58,6 +59,10 @@ class AdminJobCategoryController extends Controller
     
     public function delete($id)
     {
+        $check = Job::where('job_category_id', $id)->count();
+        if($check > 0) {
+            return redirect()->back()->with('error', 'You can not delete this item, because this is used in another place.');
+        }
         JobCategory::where('id', $id)->delete();
         return redirect()->route('admin_job_category')->with('success', 'Data is deleted successfully.');
     }
