+++ app/Http/Controllers/Admin/AdminJobExperienceController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\JobExperience;
+use App\Models\Job;
 
 class AdminJobExperienceController extends Controller
 {
@@ -54,6 +55,10 @@ class AdminJobExperienceController extends Controller
     
     public function delete($id)
     {
+        $check = Job::where('job_experience_id', $id)->count();
+        if($check > 0) {
+            return redirect()->back()->with('error', 'You can not delete this item, because this is used in another place.');
+        }
         JobExperience::where('id', $id)->delete();
         return redirect()->route('admin_job_experience')->with('success', 'Data is deleted successfully.');
     }
