+++ app/Http/Controllers/Admin/AdminJobSalaryRangeController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\JobSalaryRange;
+use App\Models\Job;
 
 class AdminJobSalaryRangeController extends Controller
 {
@@ -54,6 +55,10 @@ class AdminJobSalaryRangeController extends Controller
     
     public function delete($id)
     {
+        $check = Job::where('job_salary_range_id', $id)->count();
+        if($check > 0) {
+            return redirect()->back()->with('error', 'You can not delete this item, because this is used in another place.');
+        }
         JobSalaryRange::where('id', $id)->delete();
         return redirect()->route('admin_job_salary_range')->with('success', 'Data is deleted successfully.');
     }
