+++ app/Http/Controllers/Admin/AdminPackageController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Package;
+use App\Models\Order;
 
 class AdminPackageController extends Controller
 {
@@ -83,6 +84,10 @@ class AdminPackageController extends Controller
     
     public function delete($id)
     {
+        $check = Order::where('package_id', $id)->count();
+        if($check > 0) {
+            return redirect()->back()->with('error', 'You can not delete this item, because this is used in another place.');
+        }
         Package::where('id', $id)->delete();
         return redirect()->route('admin_package')->with('success', 'Data is deleted successfully.');
     }
